requirements.txt
azure-monitor-opentelemetry
azure-functions
azure-eventhub
azure-cosmos
pydantic
retry
aiohttp